norma passing
