+++ Makefile
@@ -46,7 +46,8 @@ SRC =	_minishell_/minishell.c \
 		executer/processes.c \
 		executer/access_utils.c \
 		executer/heredoc.c \
-		builtins/exit.c
+		builtins/exit.c \
+		builtins/builtins_utils.c
 
 OBJ = $(SRC:.c=.o)
