+++ _minishell_/minishell.c
@@ -6,7 +6,7 @@
 /*   By: asoler <asoler@student.42sp.org.br>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/05 22:57:17 by asoler            #+#    #+#             */
-/*   Updated: 2022/11/27 22:25:52 by asoler           ###   ########.fr       */
+/*   Updated: 2022/12/03 23:50:51 by asoler           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,10 +29,12 @@ int	main(int argc, char **argv, char **envp)
 // [TODO]
 // - testar variaveis vazias
 // - export deve funcionar unicamente com formato key=value
-		// - quando usado com pipe retorna menssagem | tmbm passando so = ou 1=2 
+		// - quando usado com pipe retorna menssagem
+			// - tmbm passando numero como nome da var
+			// export 1=sdfsd
+			// bash: export: `1=sdfsd': not a valid identifier
 			// export HH+JJ | ls
 			// bash: export: `HH+JJ': not a valid identifier
-			// avaliações  mini_merge  pipeline  pipex_adapt  printa_struc  print_data_struct  projetos_fase1
 // - unset -> is not working
 // // try these:
 		// export PATH=$PATH:/something/bin
